+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "homepage": "https://mmarsden89.github.io/national-park-prototype",
   "dependencies": {
+    "axios": "^0.19.0",
     "gh-pages": "^2.1.1",
     "node-sass": "^4.12.0",
     "react": "^16.9.0",
